+++ uui-core/src/hooks/useVirtualList/utils.ts
